requirements.txt
streamlit>=1.34.0
pydantic>=2.6.0
python-dotenv>=1.0.1
numpy>=1.24.0
pandas>=2.1.0
yfinance>=0.2.40
matplotlib>=3.8.0
requests>=2.31.0
tenacity>=8.2.3
python-dateutil>=2.9.0.post0
openai>=1.30.0
